% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{ccmatch}
\alias{ccmatch}
\alias{ccmatch-package}
\title{ccmatch}
\description{
R package for optimal matching of cases to controls using network flow theory.
}
\author{
Masahiro Kanai
}
\references{
http://dx.doi.org/10.2307/2290079
}

