% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ccmatch}
\alias{ccmatch}
\title{Conduct optimal matching of cases to controls using network flow theory.}
\usage{
ccmatch(x.case, x.control, n = 1, method = "Euclidean", ...)
}
\arguments{
\item{x.case}{a numeric \code{vector}.}

\item{x.control}{a numeric \code{vector}.}

\item{n}{a ratio of case:control.}

\item{method}{a function, a registry entry, or a mnemonic string referencing the proximity measure passed to \code{proxy::dist}.}

\item{...}{further arguments passed to \code{proxy::dist}.}
}
\description{
Conduct optimal matching of cases to controls using network flow theory.
}
\references{
http://dx.doi.org/10.2307/2290079
}

