% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccmatch.R
\name{ccmatch}
\alias{ccmatch}
\title{Conduct optimal matching of cases to controls using network flow theory.}
\usage{
ccmatch(x.case, x.control, n = 1, display_progress = TRUE,
  method = "Euclidean", ...)
}
\arguments{
\item{x.case}{a numeric \code{matrix}.}

\item{x.control}{a numeric \code{matrix}.}

\item{n}{a ratio of case:control.}

\item{display_progress}{a logical indicating whether to display a progress bar.}

\item{method}{a function, a registry entry, or a mnemonic string referencing the proximity measure passed to \code{proxy::dist}.}

\item{...}{further arguments passed to \code{proxy::dist}.}
}
\description{
Conduct optimal matching of cases to controls using network flow theory.
}
\references{
http://dx.doi.org/10.2307/2290079
}

